+++ src/test/hello/ApplicationTest.java
@@ -1,0 +1,16 @@
+package hello;
+
+import org.junit.Test;
+
+import static org.junit.Assert.*;
+
+/**
+ * Created by gl42ben on 2/1/17.
+ */
+public class ApplicationTest {
+    @Test
+    public void main() throws Exception {
+
+    }
+
+}
