--- src/test/java/com/example/DemoApplicationTests.java
@@ -1,18 +1,0 @@
-package com.example;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.test.context.web.WebAppConfiguration;
-import org.springframework.boot.test.SpringApplicationConfiguration;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-
-@RunWith(SpringJUnit4ClassRunner.class)
-@SpringApplicationConfiguration(classes = Application.class)
-@WebAppConfiguration
-public class DemoApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
